requirements.txt
disnake==2.*
python-dotenv==0.*
speedtest-cli==2.*
